+++ src/components/routes/dashboard/CreatorCountSpotlight.tsx
@@ -1,7 +1,7 @@
 import {useEffect, useState} from 'react'
 import {Person} from '../../../rdf/types/Person'
 import {Link} from 'react-router-dom'
-import {PAROLANGELO_ROUTE} from '../Parolangelo'
+import {PAROLANGELO_ROUTE} from '../parolangelo/Parolangelo'
 import Spotlight from '../../common/story/Spotlight'
 
 /**