+++ src/components/routes/parolangelo/ConceptLayout.tsx
@@ -1,14 +1,14 @@
 import React, {useEffect, useState} from 'react'
-import {NamedSection} from '../common/NamedSection'
+import {NamedSection} from '../../common/NamedSection'
 import {useParams} from 'react-router-dom'
-import {Concept} from '../../rdf/types/Concept'
-import {RDFStore} from '../../rdf/RDFStore'
-import {vocang} from '../../rdf/prefixes'
-import Wrapper from '../common/story/Wrapper'
-import {List} from '../common/List'
-import {Person} from '../../rdf/types/Person'
-import ConditionalComponent from '../common/conditional/ConditionalComponent'
-import InnerWrapper from '../common/story/InnerWrapper'
+import {Concept} from '../../../rdf/types/Concept'
+import {RDFStore} from '../../../rdf/RDFStore'
+import {vocang} from '../../../rdf/prefixes'
+import Wrapper from '../../common/story/Wrapper'
+import {List} from '../../common/List'
+import {Person} from '../../../rdf/types/Person'
+import ConditionalComponent from '../../common/conditional/ConditionalComponent'
+import InnerWrapper from '../../common/story/InnerWrapper'
 
 export function ConceptLayout() {